backend/requirements.txt
asgiref==3.8.1
blinker==1.9.0
click==8.1.7
Django==5.1.3
Flask==3.0.3
Flask-Cors==5.0.0
image==1.5.33
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==3.0.2
numpy==2.1.3
pillow==11.0.0
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sqlparse==0.5.2
threadpoolctl==3.5.0
Werkzeug==3.1.3
